TINY-8310: Additional cleanup, work to allow overriding and removed duplication
