+++ modules/tinymce/src/core/main/ts/api/OptionTypes.ts
@@ -255,6 +255,7 @@ export interface EditorOptions extends NormalizedEditorOptions {
   contextmenu: string[];
   font_css: string[];
   forced_root_block: string;
+  forced_root_block_attrs: Record<string, string>;
   noneditable_regexp: RegExp[];
   object_resizing?: string;
   preview_styles?: string;
