+++ modules/tinymce/src/core/main/ts/delete/MergeText.ts
@@ -5,30 +5,13 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Arr, Strings, Unicode } from '@ephox/katamari';
+import { Strings } from '@ephox/katamari';
 import { PredicateFind, Remove, SugarElement } from '@ephox/sugar';
 
 import CaretPosition from '../caret/CaretPosition';
 import * as ElementType from '../dom/ElementType';
 import * as Nbsps from '../keyboard/Nbsps';
-import { isNbsp, isWhiteSpace } from '../text/CharType';
-
-const normalizeContent = (content: string, isStartOfContent: boolean, isEndOfContent: boolean): string => {
-  const result = Arr.foldl(content, (acc, c) => {
-    // Are we dealing with a char other than some collapsible whitespace or nbsp? if so then just use it as is
-    if (isWhiteSpace(c) || isNbsp(c)) {
-      if (acc.previousCharIsSpace || (acc.str === '' && isStartOfContent) || (acc.str.length === content.length - 1 && isEndOfContent)) {
-        return { previousCharIsSpace: false, str: acc.str + Unicode.nbsp };
-      } else {
-        return { previousCharIsSpace: true, str: acc.str + ' ' };
-      }
-    } else {
-      return { previousCharIsSpace: false, str: acc.str + c };
-    }
-  }, { previousCharIsSpace: false, str: '' });
-
-  return result.str;
-};
+import * as Whitespace from '../text/Whitespace';
 
 const normalize = (node: Text, offset: number, count: number): void => {
   if (count === 0) {
@@ -45,7 +28,7 @@ const normalize = (node: Text, offset: number, count: number): void => {
   const isStartOfContent = offset === 0 && Nbsps.needsToBeNbspLeft(root, CaretPosition(node, 0));
 
   // Replace the original whitespace with the normalized whitespace content
-  node.replaceData(offset, count, normalizeContent(whitespace, isStartOfContent, isEndOfContent));
+  node.replaceData(offset, count, Whitespace.normalize(whitespace, 4, isStartOfContent, isEndOfContent));
 };
 
 const normalizeWhitespaceAfter = (node: Text, offset: number): void => {
