+++ modules/tinymce/src/core/main/ts/paste/Newlines.ts
@@ -5,17 +5,12 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Obj } from '@ephox/katamari';
+import { Arr, Obj } from '@ephox/katamari';
 
-import Entities from '../../api/html/Entities';
-import Tools from '../../api/util/Tools';
+import Entities from '../api/html/Entities';
 
-export interface RootAttrs {[key: string]: string }
-
-/**
- * Newlines class contains utilities to convert newlines (\n or \r\n) tp BRs or to a combination of the specified block element and BRs
- *
- * @private
+/*
+ * This module contains utilities to convert newlines (\n or \r\n) to BRs or to a combination of the specified block element and BRs
  */
 
 const isPlainText = (text: string): boolean => {
@@ -28,30 +23,23 @@ const isPlainText = (text: string): boolean => {
 const toBRs = (text: string): string =>
   text.replace(/\r?\n/g, '<br>');
 
-const openContainer = (rootTag: string, rootAttrs: RootAttrs): string => {
-  const attrs: string[] = [];
+const openContainer = (rootTag: string, rootAttrs: Record<string, string>): string => {
   let tag = '<' + rootTag;
 
-  if (typeof rootAttrs === 'object') {
-    for (const key in rootAttrs) {
-      if (Obj.has(rootAttrs, key)) {
-        attrs.push(key + '="' + Entities.encodeAllRaw(rootAttrs[key]) + '"');
-      }
-    }
-
-    if (attrs.length) {
-      tag += ' ' + attrs.join(' ');
-    }
+  const attrs = Obj.mapToArray(rootAttrs, (value, key) => key + '="' + Entities.encodeAllRaw(value) + '"');
+  if (attrs.length) {
+    tag += ' ' + attrs.join(' ');
   }
+
   return tag + '>';
 };
 
-const toBlockElements = (text: string, rootTag: string, rootAttrs: RootAttrs): string => {
+const toBlockElements = (text: string, rootTag: string, rootAttrs: Record<string, string>): string => {
   const blocks = text.split(/\n\n/);
   const tagOpen = openContainer(rootTag, rootAttrs);
   const tagClose = '</' + rootTag + '>';
 
-  const paragraphs = Tools.map(blocks, (p) => {
+  const paragraphs = Arr.map(blocks, (p) => {
     return p.split(/\n/).join('<br />');
   });
 
@@ -59,11 +47,11 @@ const toBlockElements = (text: string, rootTag: string, rootAttrs: RootAttrs): s
     return tagOpen + p + tagClose;
   };
 
-  return paragraphs.length === 1 ? paragraphs[0] : Tools.map(paragraphs, stitch).join('');
+  return paragraphs.length === 1 ? paragraphs[0] : Arr.map(paragraphs, stitch).join('');
 };
 
-const convert = (text: string, rootTag: string | boolean, rootAttrs: RootAttrs): string =>
-  rootTag ? toBlockElements(text, rootTag === true ? 'p' : rootTag, rootAttrs) : toBRs(text);
+const convert = (text: string, rootTag: string, rootAttrs: Record<string, string>): string =>
+  rootTag ? toBlockElements(text, rootTag, rootAttrs) : toBRs(text);
 
 export {
   isPlainText,
