+++ modules/tinymce/src/core/main/ts/paste/PasteBin.ts
@@ -5,11 +5,10 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Cell } from '@ephox/katamari';
+import { Arr, Cell } from '@ephox/katamari';
 
-import Editor from '../../api/Editor';
-import Env from '../../api/Env';
-import Tools from '../../api/util/Tools';
+import Editor from '../api/Editor';
+import Env from '../api/Env';
 
 interface PasteBin {
   readonly create: () => void;
@@ -17,22 +16,23 @@ interface PasteBin {
   readonly getEl: () => HTMLElement | null;
   readonly getHtml: () => string;
   readonly getLastRng: () => Range | null;
-  readonly isDefault: () => boolean;
-  readonly isDefaultContent: (content: any) => boolean;
 }
 
-/**
+const pasteBinDefaultContent = '%MCEPASTEBIN%';
+
+/*
  * Creates a paste bin element as close as possible to the current caret location and places the focus inside that element
  * so that when the real paste event occurs the contents gets inserted into this element
  * instead of the current editor selection element.
  */
-const create = (editor: Editor, lastRngCell: Cell<Range | null>, pasteBinDefaultContent: string): void => {
-  const dom = editor.dom, body = editor.getBody();
+const create = (editor: Editor, lastRngCell: Cell<Range | null>): void => {
+  const { dom, selection } = editor;
+  const body = editor.getBody();
 
-  lastRngCell.set(editor.selection.getRng());
+  lastRngCell.set(selection.getRng());
 
   // Create a pastebin
-  const pasteBinElm = editor.dom.add(editor.getBody(), 'div', {
+  const pasteBinElm = dom.add(editor.getBody(), 'div', {
     'id': 'mcepastebin',
     'class': 'mce-pastebin',
     'contentEditable': true,
@@ -41,8 +41,7 @@ const create = (editor: Editor, lastRngCell: Cell<Range | null>, pasteBinDefault
   }, pasteBinDefaultContent);
 
   // Move paste bin out of sight since the controlSelection rect gets displayed otherwise on Gecko
-  const browser = Env.browser;
-  if (browser.isFirefox()) {
+  if (Env.browser.isFirefox()) {
     dom.setStyle(pasteBinElm, 'left', dom.getStyle(body, 'direction', true) === 'rtl' ? 0xFFFF : -0xFFFF);
   }
 
@@ -52,13 +51,14 @@ const create = (editor: Editor, lastRngCell: Cell<Range | null>, pasteBinDefault
   });
 
   pasteBinElm.focus();
-  editor.selection.select(pasteBinElm, true);
+  selection.select(pasteBinElm, true);
 };
 
-/**
+/*
  * Removes the paste bin if it exists.
  */
 const remove = (editor: Editor, lastRngCell: Cell<Range | null>) => {
+  const dom = editor.dom;
   if (getEl(editor)) {
     let pasteBinClone;
     const lastRng = lastRngCell.get();
@@ -66,9 +66,9 @@ const remove = (editor: Editor, lastRngCell: Cell<Range | null>) => {
     // WebKit/Blink might clone the div so
     // lets make sure we remove all clones
     // TODO: Man o man is this ugly. WebKit is the new IE! Remove this if they ever fix it!
-    while ((pasteBinClone = editor.dom.get('mcepastebin'))) {
-      editor.dom.remove(pasteBinClone);
-      editor.dom.unbind(pasteBinClone);
+    while ((pasteBinClone = getEl(editor))) {
+      dom.remove(pasteBinClone);
+      dom.unbind(pasteBinClone);
     }
 
     if (lastRng) {
@@ -82,29 +82,28 @@ const remove = (editor: Editor, lastRngCell: Cell<Range | null>) => {
 const getEl = (editor: Editor): HTMLElement | null =>
   editor.dom.get('mcepastebin');
 
-/**
+const isPasteBin = (elm: Node | null): elm is HTMLElement =>
+  elm && (elm as HTMLElement).id === 'mcepastebin';
+
+/*
  * Returns the contents of the paste bin as a HTML string.
- *
- * @return {String} Get the contents of the paste bin.
  */
 const getHtml = (editor: Editor): string => {
+  const dom = editor.dom;
   // Since WebKit/Chrome might clone the paste bin when pasting
   // for example: <img style="float: right"> we need to check if any of them contains some useful html.
   // TODO: Man o man is this ugly. WebKit is the new IE! Remove this if they ever fix it!
 
   const copyAndRemove = (toElm: HTMLElement, fromElm: HTMLElement) => {
     toElm.appendChild(fromElm);
-    editor.dom.remove(fromElm, true); // remove, but keep children
+    dom.remove(fromElm, true); // remove, but keep children
   };
 
   // find only top level elements (there might be more nested inside them as well, see TINY-1162)
-  const pasteBinClones = Tools.grep(editor.getBody().childNodes, (elm) => {
-    return (elm as HTMLElement).id === 'mcepastebin';
-  }) as HTMLElement[];
-  const pasteBinElm = pasteBinClones.shift();
+  const [ pasteBinElm, ...pasteBinClones ] = Arr.filter(editor.getBody().childNodes, isPasteBin);
 
   // if clones were found, move their content into the first bin
-  Tools.each(pasteBinClones, (pasteBinClone) => {
+  Arr.each(pasteBinClones, (pasteBinClone) => {
     copyAndRemove(pasteBinElm, pasteBinClone);
   });
 
@@ -112,9 +111,9 @@ const getHtml = (editor: Editor): string => {
   // paste bin (with styles and attributes) and uses it as a default  wrapper for
   // the chunks of the content, here we cycle over the whole paste bin and replace
   // those wrappers with a basic div
-  const dirtyWrappers = editor.dom.select('div[id=mcepastebin]', pasteBinElm);
+  const dirtyWrappers = dom.select('div[id=mcepastebin]', pasteBinElm);
   for (let i = dirtyWrappers.length - 1; i >= 0; i--) {
-    const cleanWrapper = editor.dom.create('div');
+    const cleanWrapper = dom.create('div');
     pasteBinElm.insertBefore(cleanWrapper, dirtyWrappers[i]);
     copyAndRemove(cleanWrapper, dirtyWrappers[i]);
   }
@@ -122,36 +121,22 @@ const getHtml = (editor: Editor): string => {
   return pasteBinElm ? pasteBinElm.innerHTML : '';
 };
 
-const isDefaultContent = (pasteBinDefaultContent: string, content: string): boolean =>
+const isDefaultPasteBinContent = (content: string): boolean =>
   content === pasteBinDefaultContent;
-
-const isPasteBin = (elm: Element | null): boolean =>
-  elm && elm.id === 'mcepastebin';
-
-const isDefault = (editor: Editor, pasteBinDefaultContent: string): boolean => {
-  const pasteBinElm = getEl(editor);
-  return isPasteBin(pasteBinElm) && isDefaultContent(pasteBinDefaultContent, pasteBinElm.innerHTML);
-};
-
-/**
- * @private
- */
 
 const PasteBin = (editor: Editor): PasteBin => {
   const lastRng = Cell(null);
-  const pasteBinDefaultContent = '%MCEPASTEBIN%';
 
   return {
-    create: () => create(editor, lastRng, pasteBinDefaultContent),
+    create: () => create(editor, lastRng),
     remove: () => remove(editor, lastRng),
     getEl: () => getEl(editor),
     getHtml: () => getHtml(editor),
-    getLastRng: lastRng.get,
-    isDefault: () => isDefault(editor, pasteBinDefaultContent),
-    isDefaultContent: (content) => isDefaultContent(pasteBinDefaultContent, content)
+    getLastRng: lastRng.get
   };
 };
 
 export {
-  PasteBin
+  PasteBin,
+  isDefaultPasteBinContent
 };
