+++ modules/tinymce/src/core/test/ts/browser/paste/SmartPasteTest.ts
@@ -4,7 +4,7 @@ import { LegacyUnit, TinyAssertions, TinyHooks } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
-import * as SmartPaste from 'tinymce/core/paste/core/SmartPaste';
+import * as SmartPaste from 'tinymce/core/paste/SmartPaste';
 
 // Test cases for TINY-4523 - image url/anchor link paste smartpaste/pasteAsText interactions
 // Pasting an image anchor link (<a href=”….jpg”>):
